refactor diskLRU
